+++ spark-screenshot-testing/src/test/kotlin/com/adevinta/spark/tokens/ColorsScreenshot.kt
@@ -59,7 +59,6 @@ internal class ColorsScreenshot {
                     listOf(::main, ::mainContainer, ::mainVariant),
                     listOf(::support, ::supportContainer, ::supportVariant),
                     listOf(::accent, ::accentContainer, ::accentVariant),
-                    listOf(::basic, ::basicContainer),
                 ),
                 listOf(
                     listOf(::success, ::successContainer),