+++ spark/src/main/kotlin/com/adevinta/spark/components/IntentColor.kt
@@ -36,13 +36,18 @@ internal enum class IntentColors {
     /**
      * Used to match default color of such UI controls as toggles, Slider, etc.
      */
+    @Deprecated(
+        message = "Basic has been removed with the rebranding, use Support instead",
+        replaceWith = ReplaceWith("IntentColors.Support"),
+        level = DeprecationLevel.ERROR,
+    )
     Basic {
         @Composable
         override fun colors() = IntentColor(
-            color = SparkTheme.colors.basic,
-            onColor = SparkTheme.colors.onBasic,
-            containerColor = SparkTheme.colors.basicContainer,
-            onContainerColor = SparkTheme.colors.onBasicContainer,
+            color = SparkTheme.colors.support,
+            onColor = SparkTheme.colors.onSupport,
+            containerColor = SparkTheme.colors.supportContainer,
+            onContainerColor = SparkTheme.colors.onSupportContainer,
         )
     },