+++ spark/src/main/kotlin/com/adevinta/spark/components/iconbuttons/toggle/IconToggleButton.kt
@@ -155,7 +155,7 @@ internal fun SparkIconToggleButton(
 @Composable
 private fun IconToggleButtonPreview() {
     PreviewTheme {
-        val intent = IconButtonIntent.Basic
+        val intent = IconButtonIntent.Support
         var state by remember {
             mutableStateOf(false)
         }