+++ spark/src/main/kotlin/com/adevinta/spark/components/progressbar/Progressbar.kt
@@ -81,7 +81,7 @@ internal fun SparkProgressbar(
 public fun Progressbar(
     progress: () -> Float,
     modifier: Modifier = Modifier,
-    intent: ProgressbarIntent = ProgressbarIntent.Basic,
+    intent: ProgressbarIntent = ProgressbarIntent.Support,
     isRounded: Boolean = true,
 ) {
     SparkProgressbar(
@@ -107,7 +107,7 @@ public fun Progressbar(
 @Composable
 public fun ProgressbarIndeterminate(
     modifier: Modifier = Modifier,
-    intent: ProgressbarIntent = ProgressbarIntent.Basic,
+    intent: ProgressbarIntent = ProgressbarIntent.Support,
     isRounded: Boolean = true,
 ) {
     SparkProgressbar(