+++ spark/src/main/kotlin/com/adevinta/spark/components/tags/TagFilled.kt
@@ -43,7 +43,7 @@ import com.adevinta.spark.icons.SparkIcons
 @Composable
 public fun TagFilled(
     modifier: Modifier = Modifier,
-    intent: TagIntent = TagIntent.Basic,
+    intent: TagIntent = TagIntent.Support,
     leadingIcon: SparkIcon? = null,
     tint: Color? = null,
     atEnd: Boolean = false,
@@ -72,7 +72,7 @@ public fun TagFilled(
 public fun TagFilled(
     text: String,
     modifier: Modifier = Modifier,
-    intent: TagIntent = TagIntent.Basic,
+    intent: TagIntent = TagIntent.Support,
     leadingIcon: SparkIcon? = null,
     tint: Color? = null,
     atEnd: Boolean = false,
@@ -100,7 +100,7 @@ public fun TagFilled(
 public fun TagFilled(
     text: AnnotatedString,
     modifier: Modifier = Modifier,
-    intent: TagIntent = TagIntent.Basic,
+    intent: TagIntent = TagIntent.Support,
     leadingIcon: SparkIcon? = null,
     tint: Color? = null,
     atEnd: Boolean = false,
@@ -123,7 +123,7 @@ internal fun TagFilledPreview() {
     PreviewTheme {
         val icon = SparkIcons.Booster
         TagFilled("", leadingIcon = icon)
-        TagFilled("Tag Basic")
-        TagFilled("Tag Basic", leadingIcon = icon)
+        TagFilled("Tag Support")
+        TagFilled("Tag Support", leadingIcon = icon)
     }
 }