+++ spark/src/main/kotlin/com/adevinta/spark/components/tags/TagIntent.kt
@@ -32,9 +32,14 @@ public enum class TagIntent {
     /**
      * Used to match default color of such UI controls as toggles, Slider, etc.
      */
+    @Deprecated(
+        message = "Basic has been removed with the rebranding, use Support instead",
+        replaceWith = ReplaceWith("TagIntent.Support"),
+        level = DeprecationLevel.ERROR,
+    )
     Basic {
         @Composable
-        override fun colors(): IntentColor = IntentColors.Basic.colors()
+        override fun colors(): IntentColor = IntentColors.Support.colors()
     },
 
     /**