+++ spark/src/main/kotlin/com/adevinta/spark/components/tags/TagOutlined.kt
@@ -44,7 +44,7 @@ import com.adevinta.spark.icons.SparkIcons
 @Composable
 public fun TagOutlined(
     modifier: Modifier = Modifier,
-    intent: TagIntent = TagIntent.Basic,
+    intent: TagIntent = TagIntent.Support,
     leadingIcon: SparkIcon? = null,
     tint: Color? = null,
     atEnd: Boolean = false,
@@ -81,7 +81,7 @@ public fun TagOutlined(
 public fun TagOutlined(
     text: String,
     modifier: Modifier = Modifier,
-    intent: TagIntent = TagIntent.Basic,
+    intent: TagIntent = TagIntent.Support,
     leadingIcon: SparkIcon? = null,
     tint: Color? = null,
     atEnd: Boolean = false,
@@ -117,7 +117,7 @@ public fun TagOutlined(
 public fun TagOutlined(
     text: AnnotatedString,
     modifier: Modifier = Modifier,
-    intent: TagIntent = TagIntent.Basic,
+    intent: TagIntent = TagIntent.Support,
     leadingIcon: SparkIcon? = null,
     tint: Color? = null,
     atEnd: Boolean = false,
@@ -148,7 +148,7 @@ internal fun TagOutlinedPreview() {
     PreviewTheme {
         val icon = SparkIcons.Booster
         TagOutlined("", leadingIcon = icon)
-        TagOutlined("Tag Basic")
-        TagOutlined("Tag Basic", leadingIcon = icon)
+        TagOutlined("Tag Support")
+        TagOutlined("Tag Support", leadingIcon = icon)
     }
 }